[nemo-qml-plugin-email] Expose number of sending accounts, rename properties. Contributes to JB#43889
The 'numberOfTransmitAccounts' property has been added to allow the
number of email accounts which are able to transmit to be identified.
This is used by the email UI to make unavailable the ability to
compose an outgoing email if there aren't any accounts that can send it.

A couple of property names have also been changed to better reflect
their purpose.

1. 'canTransmitAccounts' becomes 'onlyTransmitAccounts'.
2. 'hasPersistentConnection' becomes 'persistentConnectionActive'.