[nemo-qml-plugin-email] Apply folder sync options before sync. Fixes JB#47389
The bulk of this code was written by Damien Caliste in incorporated into
buteo-sync-plugins-email:

https://git.sailfishos.org/mer-core/buteo-sync-plugins-email/commit/b979e82ad321fd4468e8c78a4e26069d92e52a7b

The changes have been moved here because not all sync paths go through
buteo, whereas all of them do go through nemo-qml-plugin-email.

The folder sync policy is stored under the "folderSyncPolicy"
accounts-qt5 value of the root service. It can take one of the following
values:

"all-folders"
"inbox-and-subfolders"
"inbox"
"follow-flags"
""

Its purpose is to specify which folders to synchronise when a general
sync is performed. This code change applies the policy before every sync
by setting the SynchronizationEnabled flag of all QMailFolders,
depending on the value set.