Provide folder list model filterable by type with additional properties
Provides a folder list model that can be filtered by type, for example
it can be used to show only the email folders. The model also exposes:

1. count: the number of folders as a property.
2. syncFolderList: a list of the folders with the sync flag set

This is needed for the settings pages, where the number of folders in
the model determines whether the folders will be displayed mixed in with
the other settings, or an a separate page of their own. The list of
folders with sync setting set is used to summarise the settings.

Previously the folder list was implemented using a FolderListModel by hiding
non-email folder delegates. However this made counting the number of
folders trickier. It's easier to have the count calculated by the model.