[nemo-qml-plugin-email] Interpret empty sync policy as "inbox". Contributes to JB#49049
If no sync policy was set for an account, it would previously have been
interpreted as "follow-flags". This would apply to any existing account,
when in practice the previous value would have been "inbox" for most
types of account.

This change sets a default value if no other value has been set.

For IMAP4, GMail and EAS it sets it to "inbox".

For anything else, including POP3, it sets it to "follow-flags" (i.e.
does nothing to the current folder sync flag values).

It also exposes methods for propagating the setting to the folder sync
flags.