[nemo-qml-plugin-email] Remove applyFolderSyncPolicy() functionality. Contributes to JB#50162
The application of the folder sync policy is now done by QMF and its
plugins, rather than being applied prior to every sync. The change was
made primarily because applying it in the UI thread caused delays, but
it also introduced some other inconsistencies that meant the policy
wasn't always adhered to.

This change essentially reverts the changes applied in
0427d4946d474eedf3ecb6e374a03e105335cc2a