[nemo-qml-plugin-email] Move attachment model has a property of Email.
With decryption on the fly, decrypted mail
structure will never be saved to the mail
store. It thus requires for part with
attachments to be linked to the decrypted
Email object and not fetched from mail
store anymore.

EmailAgent::downloadAttachment() with a
messageId is thus replaced by a new
EmailMessage::downloadAttachment().

AttachmentListModel, previously based on
a messageId, has been modified to become
a property of an EmailMessage object.